+++ angular/projects/wcs-formly/src/lib/formly-wcs-field-radio.component.ts
@@ -12,11 +12,12 @@ import { FieldType } from '@ngx-formly/core';
         [formControl]="formControl"
         [formlyAttributes]="field">
         <wcs-radio
-          *ngFor="let option of to.options | formlySelectOptions: field | async; let i = index"
+          *ngFor="let option of to.options | wcsFormlyOptions | async; let i = index"
           [id]="option.value"
           [attr.disabled]="to.disabled || option.disabled ? true : null"
           [value]="option.value"
-          [label]="option.label"></wcs-radio>
+          [label]="option.label"
+          [ngClass]="option.class"></wcs-radio>
       </wcs-radio-group>
     </formly-wcs-field-wrapper>
   `,