+++ angular/projects/wcs-formly/src/lib/pipes/wcs-formly-options.pipe.ts
@@ -1,0 +1,15 @@
+import { Pipe, PipeTransform } from '@angular/core';
+import { Observable, of as observableOf } from 'rxjs';
+
+@Pipe({
+  name: 'wcsFormlyOptions'
+})
+export class WcsFormlyOptionsPipe implements PipeTransform {
+  transform(options: any): Observable<any[]> {
+    if (!(options instanceof Observable)) {
+      options = observableOf(options);
+    }
+
+    return options;
+  }
+}